end_flag가 켜지지 않는다.

+++ philo_bonus/philo_create.c
@@ -6,7 +6,7 @@
 /*   By: jeseo <jeseo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/08 13:21:12 by jeseo             #+#    #+#             */
-/*   Updated: 2023/02/17 20:43:03 by jeseo            ###   ########.fr       */
+/*   Updated: 2023/02/18 17:29:53 by jeseo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,13 +18,13 @@ pid_t	*generate_philo(t_philo *shared)
 
 	shared->start = get_time();
 	pid = (pid_t *)ft_calloc(shared->conditions->philo_number, sizeof(pid_t));
-	//ㄴㅓㄹ가드
+	//null가드
 	while (shared->index < shared->conditions->philo_number)
 	{
-		shared->index++;
 		pid[shared->index] = fork();
 		if (pid[shared->index] == 0)
 		{
+			free(pid);
 			break ;
 		}
 		else if (pid[shared->index] < 0)
@@ -32,6 +32,7 @@ pid_t	*generate_philo(t_philo *shared)
 			write(2, "Error\nfork() error\n", 19);
 			return (NULL);
 		}
+		shared->index++;
 	}
 	return (pid);
 }
@@ -42,20 +43,27 @@ void	philo_wait(t_philo *shared, pid_t *childs)
 	int		num;
 	int		i;
 
-	i = 0;
 	num = shared->conditions->philo_number;
+	printf("필로 기다리자\n");
 	while (1)
 	{
 		pid = waitpid(-1, NULL, WNOHANG);
-		if (pid != 0)
+		//printf("pid: %d\n", pid);
+		if (pid != 0 && pid != -1)
 		{
-			printf("R U N%d\n", pid);
+			i = 0;
 			while (i < num)
 			{
 				if (pid != childs[i])
 					kill(childs[i], SIGKILL);
 				i++;
 			}
+			return ;
 		}
+		else if (pid == -1)
+		{
+			perror("waitpid error");
+		}
+		usleep(256);
 	}
 }
